Move frequency filtering down into PagedBytesIndexFieldData.
